+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "BusinessMonitor",
-  "version": "0.1.0+70",
+  "version": "0.1.0+71",
   "private": true,
   "type": "module",
   "scripts": {
